regulator: Add IDT ID95APM regulator driver
Please note that the interrupts for DC/DC's are not fully tested
yet. This is because the DC/DC fault detection is disabled in the
current chip revision.

Additionally the current limits are set to the lowest possible
value (LDO & DC/DC). This is for fault interrupt testing only and
shall be removed in the release version.

Signed-off-by: Stefan Roese <sr@denx.de>
Signed-off-by: Pierluigi Passaro <info@phoenixsoftware.it>